refactor: extract Telegram API utils from handler layer (#151)
Move _run_telegram_api_sync, send_message_sync, edit_message_sync,
send_photo_sync to src/utils/telegram_api.py so services no longer
import upward from bot/handlers/base.py. base.py re-exports for
backward compatibility.
