+++ src/utils/__init__.py
@@ -1,5 +1,5 @@
 """Utility modules for the Telegram Agent."""
 
-from . import retry, subprocess_helper, task_tracker, tool_check
+from . import retry, subprocess_helper, task_tracker, telegram_api, tool_check
 
-__all__ = ["task_tracker", "subprocess_helper", "retry", "tool_check"]
+__all__ = ["task_tracker", "subprocess_helper", "retry", "telegram_api", "tool_check"]
